requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
attrs==24.2.0
bitarray==2.9.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
ckzg==1.0.2
cryptography==43.0.0
cytoolz==0.12.3
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gspread==6.1.2
hexbytes==0.3.1
httplib2==0.22.0
idna==3.7
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
lru-dict==1.2.0
multidict==6.0.5
oauthlib==3.2.2
parsimonious==0.10.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rlp==4.0.1
rpds-py==0.20.0
rsa==4.9
toolz==0.12.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
web3==6.20.1
websockets==12.0
yarl==1.9.4
